+++ Source/IGListAdapterUpdater.m
@@ -132,6 +132,8 @@
         return;
     }
 
+    IGAssert(objectsWithDuplicateIdentifiersRemoved(fromObjects).count == fromObjects.count, @"The fromObjects has duplicate identifiers (it should already be dedupepd at this point).");
+    
     NSArray *toObjects = nil;
     if (toObjectsBlock != nil) {
         toObjects = objectsWithDuplicateIdentifiersRemoved(toObjectsBlock());
@@ -199,12 +201,12 @@
     // block executed in the first param block of -[UICollectionView performBatchUpdates:completion:]
     void (^batchUpdatesBlock)(IGListIndexSetResult *result) = ^(IGListIndexSetResult *result){
         executeUpdateBlocks();
-
+        
         self.applyingUpdateData = [self _flushCollectionView:collectionView
-                                             withDiffResult:result
-                                               batchUpdates:self.batchUpdates
-                                                fromObjects:fromObjects];
-
+                                              withDiffResult:result
+                                                batchUpdates:self.batchUpdates
+                                                 fromObjects:fromObjects];
+        
         [self _cleanStateAfterUpdates];
         [self _performBatchUpdatesItemBlockApplied];
     };
@@ -226,7 +228,11 @@
         [collectionView layoutIfNeeded];
 
         @try {
-            [delegate listAdapterUpdater:self willPerformBatchUpdatesWithCollectionView:collectionView];
+            [delegate  listAdapterUpdater:self
+willPerformBatchUpdatesWithCollectionView:collectionView
+                              fromObjects:fromObjects
+                                toObjects:toObjects
+                       listIndexSetResult:result];
             if (result.changeCount > 100 && IGListExperimentEnabled(experiments, IGListExperimentReloadDataFallback)) {
                 reloadDataFallback();
             } else if (animated) {
@@ -249,6 +255,7 @@
                   willCrashWithException:exception
                              fromObjects:fromObjects
                                toObjects:toObjects
+                              diffResult:result
                                  updates:(id)self.applyingUpdateData];
             @throw exception;
         }
@@ -294,6 +301,17 @@ void convertReloadToDeleteInsert(NSMutableIndexSet *reloads,
     }];
 }
 
+static NSArray<NSIndexPath *> *convertSectionReloadToItemUpdates(NSIndexSet *sectionReloads, UICollectionView *collectionView) {
+    NSMutableArray<NSIndexPath *> *updates = [NSMutableArray new];
+    [sectionReloads enumerateIndexesUsingBlock:^(NSUInteger sectionIndex, BOOL * _Nonnull stop) {
+        NSUInteger numberOfItems = [collectionView numberOfItemsInSection:sectionIndex];
+        for (NSUInteger itemIndex = 0; itemIndex < numberOfItems; itemIndex++) {
+            [updates addObject:[NSIndexPath indexPathForItem:itemIndex inSection:sectionIndex]];
+        }
+    }];
+    return [updates copy];
+}
+
 - (IGListBatchUpdateData *)_flushCollectionView:(UICollectionView *)collectionView
                                 withDiffResult:(IGListIndexSetResult *)diffResult
                                   batchUpdates:(IGListBatchUpdates *)batchUpdates
@@ -306,6 +324,7 @@ void convertReloadToDeleteInsert(NSMutableIndexSet *reloads,
 
     NSMutableIndexSet *inserts = [diffResult.inserts mutableCopy];
     NSMutableIndexSet *deletes = [diffResult.deletes mutableCopy];
+    NSMutableArray<NSIndexPath *> *itemUpdates = [NSMutableArray new];
     if (self.movesAsDeletesInserts) {
         for (IGListMoveIndex *move in moves) {
             [deletes addIndex:move.from];
@@ -315,9 +334,26 @@ void convertReloadToDeleteInsert(NSMutableIndexSet *reloads,
         moves = [NSSet new];
     }
 
-    // reloadSections: is unsafe to use within performBatchUpdates:, so instead convert all reloads into deletes+inserts
-    convertReloadToDeleteInsert(reloads, deletes, inserts, diffResult, fromObjects);
-
+    // Item reloads are not safe, if any section moves happened or there are inserts/deletes.
+    if (self.preferItemReloadsForSectionReloads
+        && moves.count == 0 && inserts.count == 0 && deletes.count == 0 && reloads.count > 0) {
+        [reloads enumerateIndexesUsingBlock:^(NSUInteger sectionIndex, BOOL * _Nonnull stop) {
+            NSMutableIndexSet *localIndexSet = [NSMutableIndexSet indexSetWithIndex:sectionIndex];
+            if (sectionIndex < [collectionView numberOfSections]
+                && sectionIndex < [collectionView.dataSource numberOfSectionsInCollectionView:collectionView]
+                && [collectionView numberOfItemsInSection:sectionIndex] == [collectionView.dataSource collectionView:collectionView numberOfItemsInSection:sectionIndex]) {
+                // Perfer to do item reloads instead, if the number of items in section is unchanged.
+                [itemUpdates addObjectsFromArray:convertSectionReloadToItemUpdates(localIndexSet, collectionView)];
+            } else {
+                // Otherwise, fallback to convert into delete+insert section operation.
+                convertReloadToDeleteInsert(localIndexSet, deletes, inserts, diffResult, fromObjects);
+            }
+        }];
+    } else {
+        // reloadSections: is unsafe to use within performBatchUpdates:, so instead convert all reloads into deletes+inserts
+        convertReloadToDeleteInsert(reloads, deletes, inserts, diffResult, fromObjects);
+    }
+    
     NSMutableArray<NSIndexPath *> *itemInserts = batchUpdates.itemInserts;
     NSMutableArray<NSIndexPath *> *itemDeletes = batchUpdates.itemDeletes;
     NSMutableArray<IGListMoveIndexPath *> *itemMoves = batchUpdates.itemMoves;
@@ -339,6 +375,7 @@ void convertReloadToDeleteInsert(NSMutableIndexSet *reloads,
                                                                                  moveSections:moves
                                                                              insertIndexPaths:itemInserts
                                                                              deleteIndexPaths:itemDeletes
+                                                                             updateIndexPaths:itemUpdates
                                                                                moveIndexPaths:itemMoves];
     [collectionView ig_applyBatchUpdateData:updateData];
     return updateData;
@@ -378,14 +415,7 @@ void convertReloadToDeleteInsert(NSMutableIndexSet *reloads,
 
 - (void)_queueUpdateWithCollectionViewBlock:(IGListCollectionViewBlock)collectionViewBlock {
     IGAssertMainThread();
-
-    // callers may hold weak refs and lose the collection view by the time we requeue, bail if that's the case
-//    if (collectionView == nil) {
-//        return;
-//    }
-
     __weak __typeof__(self) weakSelf = self;
-//    __weak __typeof__(collectionView) weakCollectionView = collectionView;
 
     // dispatch after a given amount of time to coalesce other updates and execute as one
     dispatch_after(dispatch_time(DISPATCH_TIME_NOW, self.coalescanceTime * NSEC_PER_SEC), dispatch_get_main_queue(), ^{