+++ Source/IGListAdapterUpdaterDelegate.h
@@ -10,6 +10,9 @@
 #import <IGListKit/IGListBatchUpdateData.h>
 
 @class IGListAdapterUpdater;
+@class IGListBatchUpdates;
+@class IGListIndexSetResult;
+@protocol IGListDiffable;
 
 NS_ASSUME_NONNULL_BEGIN
 
@@ -24,8 +27,15 @@ NS_SWIFT_NAME(ListAdapterUpdaterDelegate)
 
  @param listAdapterUpdater The adapter updater owning the transition.
  @param collectionView The collection view that will perform the batch updates.
+ @param fromObjects The items transitioned from in the batch updates, if any.
+ @param toObjects The items transitioned to in the batch updates, if any.
+ @param listIndexSetResults The diffing result of indices to be inserted/removed/updated/moved/etc.
  */
-- (void)listAdapterUpdater:(IGListAdapterUpdater *)listAdapterUpdater willPerformBatchUpdatesWithCollectionView:(UICollectionView *)collectionView;
+- (void)               listAdapterUpdater:(IGListAdapterUpdater *)listAdapterUpdater
+willPerformBatchUpdatesWithCollectionView:(UICollectionView *)collectionView
+                              fromObjects:(nullable NSArray <id<IGListDiffable>> *)fromObjects
+                                toObjects:(nullable NSArray <id<IGListDiffable>> *)toObjects
+                       listIndexSetResult:(nullable IGListIndexSetResult *)listIndexSetResults;
 
 /**
  Notifies the delegate that the updater successfully finished `-[UICollectionView performBatchUpdates:completion:]`.
@@ -131,6 +141,7 @@ NS_SWIFT_NAME(ListAdapterUpdaterDelegate)
  @param exception The exception thrown by the collection view.
  @param fromObjects The items transitioned from in the diff, if any.
  @param toObjects The items transitioned to in the diff, if any.
+ @param diffResult The diff result that were computed from `fromObjects` and `toObjects`.
  @param updates The batch updates that were applied to the collection view.
  */
 - (void)listAdapterUpdater:(IGListAdapterUpdater *)listAdapterUpdater
@@ -138,6 +149,7 @@ NS_SWIFT_NAME(ListAdapterUpdaterDelegate)
     willCrashWithException:(NSException *)exception
                fromObjects:(nullable NSArray *)fromObjects
                  toObjects:(nullable NSArray *)toObjects
+                diffResult:(IGListIndexSetResult *)diffResult
                    updates:(IGListBatchUpdateData *)updates;
 
 @end