+++ remodel-plugin/src/__tests__/plugins/iglistdiffable-test.ts
@@ -21,6 +21,7 @@ const ObjectSpecPlugin = IGListDiffable.createPlugin();
 
 function igListDiffableIsEqualMethod():ObjC.Method {
   return {
+    preprocessors:[],
     belongsToProtocol:Maybe.Just('IGListDiffable'),
     code: ['return [self isEqual:object];'],
     comments:[],
@@ -47,6 +48,7 @@ function igListDiffableIsEqualMethod():ObjC.Method {
 
 function igListDiffableDiffIdentifierMethodWithCode(code:string):ObjC.Method {
   return {
+    preprocessors:[],
     belongsToProtocol:Maybe.Just<string>('IGListDiffable'),
     code: [code],
     comments:[],
@@ -162,7 +164,7 @@ describe('ObjectSpecPlugins.IGListDiffable', function() {
 
         const expectedInstanceMethods:ObjC.Method[] = [
           igListDiffableIsEqualMethod(),
-          igListDiffableDiffIdentifierMethodWithCode('return [NSString stringWithFormat:@"%lld", (long long)_age];')
+          igListDiffableDiffIdentifierMethodWithCode('return @(_age);')
         ];
 
         expect(instanceMethods).toEqualJSON(expectedInstanceMethods);
@@ -201,7 +203,7 @@ describe('ObjectSpecPlugins.IGListDiffable', function() {
 
         const expectedInstanceMethods:ObjC.Method[] = [
           igListDiffableIsEqualMethod(),
-          igListDiffableDiffIdentifierMethodWithCode('return NSStringFromCGRect(_rect);')
+          igListDiffableDiffIdentifierMethodWithCode('return [NSValue valueWithCGRect:_rect];')
         ];
 
         expect(instanceMethods).toEqualJSON(expectedInstanceMethods);
@@ -254,7 +256,7 @@ describe('ObjectSpecPlugins.IGListDiffable', function() {
 
         const expectedInstanceMethods:ObjC.Method[] = [
           igListDiffableIsEqualMethod(),
-          igListDiffableDiffIdentifierMethodWithCode('return [NSString stringWithFormat:@"%lld", (long long)_age];')
+          igListDiffableDiffIdentifierMethodWithCode('return @(_age);')
         ];
 
         expect(instanceMethods).toEqualJSON(expectedInstanceMethods);